Checked in a new file, glissade_velo.F90.  This module will eventually contain many of the subroutines used to compute velocities using the Dukowicz et al. variational method.  Just a couple of subroutines so far, but will have lots more later.
Note that Makefile.am has been modified, so the code must be reconfigured to
compile the new module.

The idea moving forward is to modify the existing glissade.F90 so that it calls
the new velo routines for the SIA, SSA, and Blatter/Pattyn approximations.  Then
the mass and tracers (including temperature) will be remapped, and then the
new temperatures will be computed in each column (without any horizontal dependence).
I will get rid of glissade_remap, which is now redundant given the new remap_advection
module, and I will add glissade_thck and glissade_temp.  Should keep me busy for
a while.