Because higher order computation potentially requires input fields with three dimensions, this commit makes it an error to use a different number of vertical layers in the input netcdf file than is specified in the config file.
Currently I do not check that the levels are spaced correctly, mainly because the spacing
specification in CISM is complex and I do not want to place the burden of setting that
up on anyone creating a data file.  Also, floating point roundoff has previously proven
to be a problem when loading vertical levels.

This patch should not affect configurations with no vertical dimension in the input, if it
does this should be considered a bug, please report it to Tim.