Numerous changes to make PARDISO-4.0 callable. This is a set of sparse linear solvers (iterative and direct) that scales well on SMP architectures. configure.in altered to check for library. Makes certain a blas library is specified. The rest of the work consists of adding a new PARDISO sparse solver, which is a module with subroutines matching those in glimmer_sparse_slap.F90. Many fields have to be specified for PARDISO's working arrays, and input variables. Also wrote a triad to compressed (sorted) row routine. At this point it compiles, but is untested.
TODO: Complete the transfer of coefient matrix and rhs to PARDISO data types and
begin testing. Recheck all PARDISO solver setting and make certain they are
appropriate for the unsymmetric, sparse systems that we have. Begin looking into
memory allocation proceedures for the purpose of smoothing the implementation of
OpenAD.