Added configure option --enable-mpi.  This is to separate using Trilinos and using mpi.  Setting --enable-mpi=yes activates the HAVE_MPI flag, adds the mpi flags to CXXFLAGS, FCFLAGS, and LDFLAGS, and finally it sets the Trilinos library list to match a parallel build.  Using Coyote modules, mpi options are stored in the environment variables MPI_COMPILE_FLAGS and MPI_LD_FLAGS.  Configure.in uses these variable names.  This can be modified later.  Skipping this flag or setting --enable-mpi=no sets the Trilinos library list to match a serial build.  The Trilinos library list is added to LDFLAGS when --with-trilinos is set.
The current implementation could use some clean up, but it works for now without users having to manually modify configure.in or Makefile to switch between running Trilinos in serial or parallel.  The list of Trilinos libraries was copied from my Trilinos builds.  The "include makefile.external" does not worry the same with autoconf.  Still working on this.