Modified the interface to the Trilinos Linear solver.
These changes are to make get the interface closer to
what will be needed by the distributed branch where the
matrix is computed in parallel. Instead of Trilinos 
linear solvers deciding on the parallel partitioning, 
the partitioning is now given to Trilinos. 

For now, I created a function to do the partitioning in
the same way we had previously done, so the code
still works the same as before this commit. 

All changes are in the "cpp" directory or withing
  ifdef TRILINOS in glam_strs2.F90, so no one else
should be effected by these changes.