Major cleanup of everything to do with Trilinos. **This should not change any of the algorithms.
Now that the implementation has stabilized, I went 
through and renamed most functions so that they
are more descriptive. I also renamed the files,
and cleaned up some implementation. I removed
several files that were no longer used and
deleted lots of commented-out lines of code. 
I added documentation that makes more sense 
to me, some of it in a new README file in src/cpp.

I changed the build process so that the cpp
files are collected in a library, and that
library is added to the link line for the
executables. Previously, the *.o files were
added direcly to libglimmer.a. I also removed
a line that had been unnecessarily adding cpp 
include paths (-I../cpp) to fortran compiles.

I verified that my hump and gis_20km problems
work the same as before, both --with-trilinos
and without, and with both the "sparse_easy_solve"
and "load_directly_into_trilinos" implementations. 
I apologize if I somehow broke the build on
other platforms.