Cleaned up the use of Trilinos in glimmer:
(1) Separated the two implementations of the 
Trilinos solver using the which_ho_sparse flag. 
* which_ho_sparse=3  uses Trilinos through 
    sparese_easy_solve and the Triad format.
* which_ho_sparse=4  calls what I called 
    "STANDALONE_TRILINOS_SOLVER" which bypasses
     sparse_easy_solve and the Triad matrix
     and loads directly into the Trilinos format.

(2) Added the STANDALONE_TRILINOS_SOLVER option
under the JFNK algorithm. This required some
invasive changes in the cpp directory to store
and recall different matrices (which JF refers
to as the A and C matrices), and to perform
matrix multiply. JF: I changed res_vect a
little but it gives the ame result.

(3) Cosmetic chages: All Trilinos options
in glam_strs2.F90 and glimmer_sparse.F90 were
leading to unsightly code with blocks inside
  #ifdef TRILINOS. I removed all these #ifdefs
from these source files. Instead, I moved all 
that logic out of the way and into
glimmer_sparse_trilinos.F90.  So, all the 
subroutines having to do with Trilinos are now
always compiled in. HOWEVER, if you compile 
without -DTRILINOS the contents of these 
subroutines are just something like:
 call write_log("You need Trilinos for this", GM_FATAL);

** My test problems run the same as before, 
both compiled with and without Trilinos.