Fixed problem at upper edge of staggered boundary in parallel_single version.
We are focusing on being certain that the serial version runs correctly.  
We found that the upper edge of the staggered grid was not being processed by indxvelostr() so it was not setting these cells to zero.  If they happened to initialize to a non-zero value, then they were being treated as having ice later in the program.  

Fixed by setting staggered_uhalo to 0 rather than 1 in parallel_single.F90.  The size of the staggered array is queried to set the upper bound of the do loops in indxvelostr(), so we don't need to subtract 1.