Parallel Trilinos
Using parallel_single.F90, one can now execute parallel Trilinos 
runs with number of processors > 1.  This is in the style of the LANL parallel branch.  
After the merge with parallel, we could only use one processor for Trilinos runs.  
We are calling this style of execution "parallel_single."  

This executes N separate processes which duplicate all calculations outside
of the Trilinos computation.  The Trilinos computation is executed in cooperation with all nodes
and the results are returned to each node. Each node contains and computes the entire grid. 

Parallel_single was originally written only with a single process in mind.  It has been 
modified to determine if number of tasks >= 2 and to adjust accordingly for the 
parallel_single, Trilinos modes. 

Jeff