Nox nonlinear solver is now working!
With Kate, we fixed the final difference between JFNK
in glam_strs2 and the NOX (Trilinos) implementation.
The convergence histories are not the same, but similar.
The norm of the final solution agrees to 9 digits.

To get good convergence for the hump problem, the
inexact Newton option had to be switched on. There
are numerous tolerances and options to tweak. For
now, NOX requires the input.xml file from the cpp
directory. 

All this should be off by default.