Distributed execution of the Picard solver.
This version has distributed execution of the Picard solver using Trilinos.  The remainder of the time step loop is executed serially.  Distributed variables that are referenced in the serial section are gathered to main_task using distributed_gather_var() defined in parallel_mpi.F90.  At the end of the time step loop in simple_glide.F90, the gathered variables are scattered back onto the processes using distributed_scatter_var().  These routines have not been added to parallel_single.F90 yet.  Be certain parallel_mpi.F90 is copied to parallel.F90 before compiling.

This version does start into the second time step.  Kate reports that there are some data issues in the gather/scatter exchange that need to be worked out.

globalIDs must be #define'd in glam_strs2.F90 for distributed execution number of processors > 1.  GlobalIDs is a numbering scheme for the variables and equations that is independent of the process.  

A configuration file, hump_mpi.PP.config, has been added to the tests/ho-other directory for running this version.

It has been tested on Jaguar.

Jeff, 1/7/11