Fixed segfault and compilation problems.
The segfault that was occuring for larger problems has been fixed.  The segfault was occuring in the residual calculation at the top of the solver loop.  Tested on hump.60 and hump.120.

Also updated parallel_slap.F90, so simple_glide can be compiled to use the SLAP library.  In such a configuration it does not require MPI. Copy parallel_slap.F90 over parallel.F90 in src/fortran.  This has been compiled successfully on a Mac using gfortran and g95.

Jeff