Parallel_single Fix and GlobalIDs
This version fixes a problem discovered by Pat Worley where the parallel_single (i.e., parallel Trilinos solver) version of the distributed program had radically fewer Picard iterations when number of processors >> 1.  This was determined to be caused by most of the nodes skipping the serial calculations because only task 0 was marked as main_task.  Further, the other nodes ignored the various barriers, scatters, and gathers, because the design assumption for parallel_single is that every processor will duplicate all calculations and therefore hold identical state, thus communication is unnecessary.  

During this testing it was also determined that GlobalIDs is not compatible with parallel_single for n>1, because we have set the Trilinos library to apportion the returned solution to that which each processor owns.  This is the correct behavior for parallel_mpi for n>=1, but incorrect for n>1 for parallel_single.

Jeff