Changed xml file for Trilinos options.
Renamed strat1.xml to trilinosOptions.xml. Not only is 
this a better name, it also is reformatted to make room 
for the nonlinear solvers options in the same file.

The formatting includes one level of nesting at the top:
<ParameterList>
  "strat1.xml" file contents
  <optional>  "input.xml" file contents for nonlinear solver
</ParameterList>

I added an error message and graceful exit() instead of 
core dump when there is a problem with the XML file.

I also added updated the content of the README file in 
libglimmer-trilinos directory.