Enable JFNK method for distributed variables
This version enables the JFNK solver method to operate on distributed variables.  The changes required were limited to glam_strs2.F90 routines.  I was able to follow the steps I had previously applied to the Picard Iteration solver.

There is one step pending in this version.  That is to determine the correct location for a staggered halo update of uvel and vvel that is analogous to the update that was required in Picard Iteration to bring it into agreement with the serial version.

Jeff