Fast hash map search replacing linear search in distributed_globalID_to_localindex()
When globalIDs are activated, the linear search in distributed_globalID_to_localindex() was determined to be an extreme bottleneck for large, parallel_single runs.  For the original ID scheme for Trilinos this was a simple identity function.  GlobalIDs are necessary for parallel execution, because each processor must refer to the equations by the same IDs for Trilinos to work correctly.  But the RHS of the equations is referred to by the local index into the solution vector, not the globalID.

M      src/cpp/matrixInterface.cpp
     For reverting to simple IDs with parallel_single the full map of the vector is required.  Was commented out for GlobalIDs parallel_mpi case.  Modified so that this is determined at runtime.

M      src/cpp/trilinosPartition.cpp
     Contains "Jeff's Quick Hash Table GlobalIDs solution."  Three functions added called globalidsadd_(), globalidsclear_(), and globalidsget_().

M      src/fortran/glam_strs2.F90
     Eliminated inefficient linear search in distributed_globalID_to_localindex() for finding corresponding RHS element for an equation specified by globalID.