added raising an exception to the logic that tries to retrieve the module name from the "end module" string and fails to do so. Superficially the problem was that Fortran does not require the module name to occur in the 'end' string and then the subsequent logic does not introduce dependencies to that module in the make rules. To catch the exception and set the return code the main logic was put in a try block and the output file if any removed as part of the exception handling.
This is only a flimsy bandaid though.  In general, the  correct scanning and parsing of Fortran is more complicated  than a simple substring match and the functionality can easily be broken again such as by having 2 spaces between "end" and "module".  
One would have to use proper regular expressions for more robust parsing.